Broke out the SubredditSortController class from MainActivity.java
